Reorganize Connector
