+++ src/Core/Connector/Commands/BeginAuthRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Smolblog\Core\Connector;
+namespace Smolblog\Core\Connector\Commands;
 
 use Smolblog\Framework\Command;
