--- src/Core/Connector/ConnectorConfig.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Smolblog\Core\Connector;
-
-/**
- * Configuration information about a Connector
- */
-class ConnectorConfig {
-	/**
-	 * Create the configuration object.
-	 *
-	 * @param string $slug URL-friendly identifier for the Connector.
-	 */
-	public function __construct(
-		public readonly string $slug
-	) {
-	}
-}
