+++ src/Core/Connector/Entities/ChannelWriter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Smolblog\Core\Connector;
+namespace Smolblog\Core\Connector\Entities;
 
 use Smolblog\Framework\Writer;
